pyproject.toml
[tool.poetry]
name = "reroot-regulation_actor_arena"
version = "0.1.0"
description = ""
authors = ["ynnck <ylenaerts89@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10.6"
Django = "^4.1"
requests = "^2.28.1"
pyhumps = "^3.7.2"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
